# docs/swagger.yaml
basePath: /api
definitions:
...
  requests.AdminUpdateUserRequest:
    properties:
      user:
        properties:
          email:
            type: string
          role:
            type: string
          username:
            type: string
        type: object
    type: object
...
  requests.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  requests.RegisterRequest:
    properties:
      user:
        properties:
          email:
            type: string
          password:
            minLength: 6
            type: string
          username:
            type: string
        required:
        - email
        - password
        - username
        type: object
    type: object
...
  requests.UpdateUserRequest:
    properties:
      user:
        properties:
          email:
            type: string
          password:
            type: string
          username:
            type: string
        type: object
    type: object
host: localhost:8080
info:
  contact: {}
  description: A RESTful API for a food delivery application
  title: Food Delivery API
  version: "1.0"
paths:
...
  /admin/users:
    get:
      description: Get all users with pagination (Admin only)
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
      security:
      - BearerAuth: []
      summary: Get all users
      tags:
      - Admin - Users
  /admin/users/{id}:
    delete:
      description: Delete a user (Admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
      security:
      - BearerAuth: []
      summary: Delete user (Admin only)
      tags:
      - Admin - Users
    put:
      consumes:
      - application/json
      description: Update any user profile (Admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User update data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.AdminUpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User updated
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
      security:
      - BearerAuth: []
      summary: Update user (Admin only)
      tags:
      - Admin - Users
...
  /user:
    get:
      description: Get the profile of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update the profile of the authenticated user
      parameters:
      - description: Updated user data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
      security:
      - BearerAuth: []
      summary: Update current user profile
      tags:
      - User
  /user/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and receive JWT token
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
      summary: Login user
      tags:
      - Auth
  /user/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: User registration data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
      summary: Register a new user
      tags:
      - Auth
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"